+++ fabric/src/main/java/dynamic_asset_generator/fabric/PlatformImpl.java
@@ -1,0 +1,11 @@
+package dynamic_asset_generator.fabric;
+
+import net.fabricmc.loader.api.FabricLoader;
+
+import java.nio.file.Path;
+
+public class PlatformImpl {
+    public static Path getConfigFolder() {
+        return FabricLoader.getInstance().getConfigDir();
+    }
+}